Safari-611.1.10/LayoutTests/accessibility/win/document-enabled-state-expected.txt

This tests that the enabled state of the document is true.

Bug #67974: MSAA: WebKit reports the document state as disabled

This test must be run inside of DumpRenderTree.
PASS accessibilityController.focusedElement.parentElement().role is "AXRole: document"
PASS accessibilityController.focusedElement.parentElement().isEnabled is true
